testing/20140821.html

<!DOCTYPE html>
<title>IRC logs: w3c / #testing / 20140821</title>
<meta name="viewport" content="width=device-width">
<style>
ol { font-family: monospace; color: #000; font-size: 90%; list-style: none; margin: 0; padding: 0; }
li { margin: 0; padding: 0; white-space: pre-wrap; word-wrap: break-word; }
li:target { background: #f90; }
li.hide { display: none; }
li.flagged { background: yellow; }
li.filtered { color: #aaa; font-size: 60%; }
li.filtered:hover { font-size: 100%; color: #000; height: auto; }
li span { display: inline; margin: 0; padding: 0 10px; background: #eee; color: #000; text-decoration: none; cursor: pointer; }
li.flagged span { background: #f00; }
li:hover span { background: #ff0 }
input#filter { margin: 0 10px; }
label { cursor: pointer; }
p span { background: #ff0; margin-left: 10px; }
</style>
<script src="/irc-logs/s.js"></script>
<h1 id="w3c/testing.20140821.log"><a href="/irc-logs/">/irc-logs</a> / w3c / <a href="irc://irc.w3.org:6665/testing">#testing</a> / 2014-08-21 / <a href="#end">end</a></h1>
<p>Options: 
<ol id="lines">
 <li id="l-1"><a href="#l-1">#</a> Session Start: Thu Aug 21 00:00:00 2014</li>
 <li id="l-2"><a href="#l-2">#</a> Session Ident: #testing</li>
 <li id="l-3"><a href="#l-3">#</a> [01:22] * Joins: MK (kawamori@public.cloak)</li>
 <li id="l-4"><a href="#l-4">#</a> [01:31] * Quits: ArtB (~abarsto@public.cloak) (&quot;Leaving.&quot;)</li>
 <li id="l-5"><a href="#l-5">#</a> [02:14] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-6"><a href="#l-6">#</a> [02:14] * Joins: MK_ (kawamori@public.cloak)</li>
 <li id="l-7"><a href="#l-7">#</a> [02:18] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-8"><a href="#l-8">#</a> [02:19] * Quits: glenn_ (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-9"><a href="#l-9">#</a> [02:22] * Joins: glenn__ (~gadams@public.cloak)</li>
 <li id="l-10"><a href="#l-10">#</a> [02:25] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-11"><a href="#l-11">#</a> [02:37] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-12"><a href="#l-12">#</a> [02:42] * Quits: glenn__ (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-13"><a href="#l-13">#</a> [02:46] * Quits: MK_ (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-14"><a href="#l-14">#</a> [03:26] * Joins: MK (kawamori@public.cloak)</li>
 <li id="l-15"><a href="#l-15">#</a> [03:28] * Joins: glenn_ (~gadams@public.cloak)</li>
 <li id="l-16"><a href="#l-16">#</a> [03:33] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-17"><a href="#l-17">#</a> [03:39] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-18"><a href="#l-18">#</a> [03:40] * Quits: ptressel (~chatzilla@public.cloak) (Client closed connection)</li>
 <li id="l-19"><a href="#l-19">#</a> [03:41] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-20"><a href="#l-20">#</a> [03:45] * Quits: glenn_ (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-21"><a href="#l-21">#</a> [03:52] * Joins: glenn_ (~gadams@public.cloak)</li>
 <li id="l-22"><a href="#l-22">#</a> [03:57] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-23"><a href="#l-23">#</a> [04:01] * Joins: MK (kawamori@public.cloak)</li>
 <li id="l-24"><a href="#l-24">#</a> [04:02] * Quits: MK (kawamori@public.cloak) (Client closed connection)</li>
 <li id="l-25"><a href="#l-25">#</a> [04:02] * Joins: MK (kawamori@public.cloak)</li>
 <li id="l-26"><a href="#l-26">#</a> [04:07] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-27"><a href="#l-27">#</a> [04:11] * Quits: glenn_ (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-28"><a href="#l-28">#</a> [04:38] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-29"><a href="#l-29">#</a> [04:39] * Joins: glenn_ (~gadams@public.cloak)</li>
 <li id="l-30"><a href="#l-30">#</a> [04:44] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-31"><a href="#l-31">#</a> [05:10] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-32"><a href="#l-32">#</a> [05:15] * Joins: glenn__ (~gadams@public.cloak)</li>
 <li id="l-33"><a href="#l-33">#</a> [05:16] * Quits: glenn_ (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-34"><a href="#l-34">#</a> [05:19] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-35"><a href="#l-35">#</a> [05:53] * Joins: zcorpan_ (~zcorpan@public.cloak)</li>
 <li id="l-36"><a href="#l-36">#</a> [06:02] * Quits: glenn__ (~gadams@public.cloak) (Client closed connection)</li>
 <li id="l-37"><a href="#l-37">#</a> [06:17] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-38"><a href="#l-38">#</a> [06:24] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-39"><a href="#l-39">#</a> [06:27] * Joins: MK (kawamori@public.cloak)</li>
 <li id="l-40"><a href="#l-40">#</a> [06:29] * Joins: ptressel (~chatzilla@public.cloak)</li>
 <li id="l-41"><a href="#l-41">#</a> [06:40] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-42"><a href="#l-42">#</a> [06:41] * Joins: MK_ (kawamori@public.cloak)</li>
 <li id="l-43"><a href="#l-43">#</a> [06:41] * Quits: MK_ (kawamori@public.cloak) (&quot;Leaving...&quot;)</li>
 <li id="l-44"><a href="#l-44">#</a> [06:47] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-45"><a href="#l-45">#</a> [07:41] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-46"><a href="#l-46">#</a> [07:48] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-47"><a href="#l-47">#</a> [08:42] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-48"><a href="#l-48">#</a> [08:49] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-49" _a="MikeSmith" _r="zcorpan_"><a href="#l-49">#</a> [08:57] &lt;MikeSmith&gt; zcorpan_: so I'm finding that the wpt html5lib test generator is coincidentally already skipping all the #script-off tests</li>
 <li id="l-50" _a="MikeSmith" _r="zcorpan_"><a href="#l-50">#</a> [08:58] &lt;MikeSmith&gt; zcorpan_: because it already skips tests that have the same hash, and all of those existing #script-off cases are duplicates of an existing test</li>
 <li id="l-51" _a="MikeSmith" _r="zcorpan_"><a href="#l-51">#</a> [08:59] &lt;MikeSmith&gt; zcorpan_: it emits already, e.g., &quot;WARNING: id c7f41e79f00db5b41872c0ef1443094e7ad5bc22 seen multiple times in file tests16 this time for test (None, &lt;!doctype html&gt;&lt;noscript&gt;&lt;!--&lt;noscript&gt; &lt;/noscript&gt;--&gt;&lt;/noscript&gt;) before for test (None, u'&lt;!doctype html&gt;&lt;noscript&gt;&lt;!--&lt;noscript&gt;&lt;/noscript&gt;--&gt;&lt;/noscript&gt;'), skipping&quot;</li>
 <li id="l-52" _a="MikeSmith" _r="zcorpan_"><a href="#l-52">#</a> [09:01] &lt;MikeSmith&gt; zcorpan_: at least I can say that if I add a continue for the #script-off condition in the loop that generates the test, I can see that the condition is getting evaluated as expected, and continuing without outputting that test</li>
 <li id="l-53" _a="zcorpan_" _r="MikeSmith"><a href="#l-53">#</a> [09:01] &lt;zcorpan_&gt; MikeSmith: interesting. still <a href="http://w3c-test.org/submissions/1204/html/syntax/parsing/html5lib_tests16.html">http://w3c-test.org/submissions/1204/html/syntax/parsing/html5lib_tests16.html</a> fails in firefox, not sure if it's a bug in the test</li>
 <li id="l-54" _a="MikeSmith" _r="zcorpan_"><a href="#l-54">#</a> [09:01] &lt;MikeSmith&gt; zcorpan_: but then when I do a git diff, there's no diff</li>
 <li id="l-55" _a="zcorpan_"><a href="#l-55">#</a> [09:02] &lt;zcorpan_&gt; expected and actual look the same to me</li>
 <li id="l-56" _a="MikeSmith" _r="zcorpan_"><a href="#l-56">#</a> [09:02] &lt;MikeSmith&gt; zcorpan_: OK I guess I'll look at just that one a bit and see what the problem is</li>
 <li id="l-57" _a="zcorpan_" _r="MikeSmith"><a href="#l-57">#</a> [09:03] &lt;zcorpan_&gt; MikeSmith: seems good to explicitly skip #script-off anyway in case future tests have different order</li>
 <li id="l-58" _a="MikeSmith" _r="yeah"><a href="#l-58">#</a> [09:03] &lt;MikeSmith&gt; yeah, agreed</li>
 <li id="l-59"><a href="#l-59">#</a> [09:03] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
 <li id="l-60" _a="MikeSmith"><a href="#l-60">#</a> [09:03] &lt;MikeSmith&gt; I will add that</li>
 <li id="l-61" _a="MikeSmith"><a href="#l-61">#</a> [09:04] &lt;MikeSmith&gt; but for now at least it's not going to fix that /html5lib_tests16.html problem</li>
 <li id="l-62" _a="MikeSmith" _r="zcorpan_"><a href="#l-62">#</a> [09:05] &lt;MikeSmith&gt; zcorpan_: you're getting just one failure in firefox, right? for 5423bb28649f37e70a0559cba78c3b253a60c277</li>
 <li id="l-63" _a="zcorpan_"><a href="#l-63">#</a> [09:05] &lt;zcorpan_&gt; yes</li>
 <li id="l-64" _a="MikeSmith"><a href="#l-64">#</a> [09:06] &lt;MikeSmith&gt; ok</li>
 <li id="l-65" _a="MikeSmith"><a href="#l-65">#</a> [09:10] &lt;MikeSmith&gt; damn it's annoying that the critic UI won't let me select and copy from diffs</li>
 <li id="l-66" _a="MikeSmith" _r="zcorpan_"><a href="#l-66">#</a> [09:18] &lt;MikeSmith&gt; zcorpan_: it seems for that test now, the Actual has a U+000A at the end, while the Expected doesn't (either that or the other way around)</li>
 <li id="l-67" _a="zcorpan_" _r="MikeSmith"><a href="#l-67">#</a> [09:19] &lt;zcorpan_&gt; MikeSmith: i think you can hold some key like ctrl, except that won't work on mac</li>
 <li id="l-68" _a="MikeSmith"><a href="#l-68">#</a> [09:20] &lt;MikeSmith&gt; oh ok, good to know. I can experiment and if the cmd key works maybe</li>
 <li id="l-69" _a="zcorpan_" _r="MikeSmith"><a href="#l-69">#</a> [09:20] &lt;zcorpan_&gt; MikeSmith: ah yeah. expected &quot;......\n&quot;</li>
 <li id="l-70"><a href="#l-70">#</a> [09:30] * Joins: dom (dom@public.cloak)</li>
 <li id="l-71" _a="heycam|away"><a href="#l-71">#</a> [09:38] * heycam|away is now known as heycam</li>
 <li id="l-72"><a href="#l-72">#</a> [09:42] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-73"><a href="#l-73">#</a> [09:50] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-74" _a="MikeSmith" _r="zcorpan_"><a href="#l-74">#</a> [09:51] &lt;MikeSmith&gt; zcorpan_: r? <a href="https://github.com/html5lib/html5lib-tests/pull/49">https://github.com/html5lib/html5lib-tests/pull/49</a> (if you have push access)</li>
 <li id="l-75" _a="zcorpan_" _r="MikeSmith"><a href="#l-75">#</a> [09:52] &lt;zcorpan_&gt; MikeSmith: r+ but no access</li>
 <li id="l-76" _a="Ms2ger"><a href="#l-76">#</a> [09:52] &lt;Ms2ger&gt; r+ and access</li>
 <li id="l-77" _a="MikeSmith"><a href="#l-77">#</a> [09:54] &lt;MikeSmith&gt; dank merci</li>
 <li id="l-78" _a="jgraham"><a href="#l-78">#</a> [09:55] * jgraham invites zcorpan_ and MikeSmith to the html5lib-test team</li>
 <li id="l-79" _a="MikeSmith"><a href="#l-79">#</a> [09:56] &lt;MikeSmith&gt; yippee ki yay, switch to XML-formatted test data files on the way</li>
 <li id="l-80" _a="jgraham"><a href="#l-80">#</a> [09:58] * jgraham sends MikeSmith to the naughty step</li>
 <li id="l-81"><a href="#l-81">#</a> [10:06] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-82" _a="MikeSmith" _r="jgraham"><a href="#l-82">#</a> [10:09] &lt;MikeSmith&gt; jgraham: hmm I get &quot;ERROR: Permission to html5lib/html5lib-python.git denied to sideshowbarker&quot; when I try to push</li>
 <li id="l-83" _a="jgraham"><a href="#l-83">#</a> [10:09] &lt;jgraham&gt; Did you accept the invitation?</li>
 <li id="l-84" _a="MikeSmith"><a href="#l-84">#</a> [10:09] &lt;MikeSmith&gt; yeah</li>
 <li id="l-85" _a="Ms2ger"><a href="#l-85">#</a> [10:10] &lt;Ms2ger&gt; html5lib-python?</li>
 <li id="l-86" _a="jgraham"><a href="#l-86">#</a> [10:10] &lt;jgraham&gt; Oh right</li>
 <li id="l-87" _a="jgraham"><a href="#l-87">#</a> [10:10] &lt;jgraham&gt; html5lib-tests</li>
 <li id="l-88" _a="MikeSmith"><a href="#l-88">#</a> [10:10] &lt;MikeSmith&gt; I was pushing to update the html5lib/tests/testdata pointer</li>
 <li id="l-89" _a="MikeSmith"><a href="#l-89">#</a> [10:11] &lt;MikeSmith&gt; can make a PR instead if that's better</li>
 <li id="l-90" _a="jgraham"><a href="#l-90">#</a> [10:19] &lt;jgraham&gt; I guess a PR is better</li>
 <li id="l-91"><a href="#l-91">#</a> [10:19] * Quits: ptressel (~chatzilla@public.cloak) (Client closed connection)</li>
 <li id="l-92" _a="MikeSmith" _r="jgraham"><a href="#l-92">#</a> [10:21] &lt;MikeSmith&gt; jgraham: hai <a href="https://github.com/html5lib/html5lib-python/pull/164">https://github.com/html5lib/html5lib-python/pull/164</a></li>
 <li id="l-93" _a="jgraham" _r="MikeSmith"><a href="#l-93">#</a> [10:23] &lt;jgraham&gt; MikeSmith: Thanks</li>
 <li id="l-94" _a="MikeSmith"><a href="#l-94">#</a> [10:24] &lt;MikeSmith&gt; どうも</li>
 <li id="l-95"><a href="#l-95">#</a> [10:25] * Joins: ptressel (~chatzilla@public.cloak)</li>
 <li id="l-96"><a href="#l-96">#</a> [10:29] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-97"><a href="#l-97">#</a> [10:29] -gitbot:#testing- [web-platform-tests] sideshowbarker pushed 1 new commit to submission/sideshowbarker/html5lib-test-update: <a href="https://github.com/w3c/web-platform-tests/commit/43929a261be068af5c5536885bf275e9843e094d">https://github.com/w3c/web-platform-tests/commit/43929a261be068af5c5536885bf275e9843e094d</a></li>
 <li id="l-98"><a href="#l-98">#</a> [10:29] -gitbot:#testing- web-platform-tests/submission/sideshowbarker/html5lib-test-update 43929a2 Michael[tm] Smith: Drop extraneous LF from expected test result.</li>
 <li id="l-99"><a href="#l-99">#</a> [10:29] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-100" _a="MikeSmith" _r="zcorpan_"><a href="#l-100">#</a> [10:31] &lt;MikeSmith&gt; zcorpan_: all passing now in firefox <a href="http://web-platform.test:8000/html/syntax/parsing/html5lib_tests16.html">http://web-platform.test:8000/html/syntax/parsing/html5lib_tests16.html</a></li>
 <li id="l-101" _a="zcorpan_" _r="MikeSmith"><a href="#l-101">#</a> [10:31] &lt;zcorpan_&gt; MikeSmith: great. is the test generator part of wpt or part of html5lib?</li>
 <li id="l-102" _a="MikeSmith" _r="zcorpan_"><a href="#l-102">#</a> [10:32] &lt;MikeSmith&gt; zcorpan_: part of wpt -- I'll do a PR for that right now</li>
 <li id="l-103" _a="zcorpan_"><a href="#l-103">#</a> [10:32] &lt;zcorpan_&gt; k</li>
 <li id="l-104"><a href="#l-104">#</a> [10:33] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-105"><a href="#l-105">#</a> [10:33] -gitbot:#testing- [web-platform-tests] zcorpan pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/0218bbda376c4bffa387a01f56c2f882cc333c7b">https://github.com/w3c/web-platform-tests/commit/0218bbda376c4bffa387a01f56c2f882cc333c7b</a></li>
 <li id="l-106"><a href="#l-106">#</a> [10:33] -gitbot:#testing- web-platform-tests/master 0218bbd Simon Pieters: Merge pull request #1204 from w3c/submission/sideshowbarker/html5lib-test-update...</li>
 <li id="l-107"><a href="#l-107">#</a> [10:33] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-108" _a="MikeSmith" _r="zcorpan_"><a href="#l-108">#</a> [10:37] &lt;MikeSmith&gt; zcorpan_: tack</li>
 <li id="l-109"><a href="#l-109">#</a> [10:37] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-110"><a href="#l-110">#</a> [10:37] -gitbot:#testing- [web-platform-tests] sideshowbarker deleted submission/sideshowbarker/html5lib-test-update at 43929a2: <a href="https://github.com/w3c/web-platform-tests/commit/43929a2">https://github.com/w3c/web-platform-tests/commit/43929a2</a></li>
 <li id="l-111"><a href="#l-111">#</a> [10:37] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-112"><a href="#l-112">#</a> [10:43] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-113"><a href="#l-113">#</a> [10:46] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-114"><a href="#l-114">#</a> [10:46] -gitbot:#testing- [web-platform-tests] sideshowbarker created submission/sideshowbarker/script-off (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/4e77ae84fb71">https://github.com/w3c/web-platform-tests/commit/4e77ae84fb71</a></li>
 <li id="l-115"><a href="#l-115">#</a> [10:46] -gitbot:#testing- web-platform-tests/submission/sideshowbarker/script-off 4e77ae8 Michael[tm] Smith: Skip html5lib script-off tests. Closes #1203.</li>
 <li id="l-116"><a href="#l-116">#</a> [10:46] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-117"><a href="#l-117">#</a> [10:46] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-118"><a href="#l-118">#</a> [10:46] -gitbot:#testing- [web-platform-tests] sideshowbarker opened pull request #1206: Skip html5lib script-off tests. Closes #1203. (master...submission/sideshowbarker/script-off) <a href="https://github.com/w3c/web-platform-tests/pull/1206">https://github.com/w3c/web-platform-tests/pull/1206</a></li>
 <li id="l-119"><a href="#l-119">#</a> [10:46] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-120"><a href="#l-120">#</a> [10:50] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-121"><a href="#l-121">#</a> [11:05] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-122"><a href="#l-122">#</a> [11:05] -gitbot:#testing- [web-platform-tests] jgraham pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/f9a52ef35dee111bd28712cbd829d7925f3cd8d8">https://github.com/w3c/web-platform-tests/commit/f9a52ef35dee111bd28712cbd829d7925f3cd8d8</a></li>
 <li id="l-123"><a href="#l-123">#</a> [11:05] -gitbot:#testing- web-platform-tests/master f9a52ef jgraham: Merge pull request #1206 from w3c/submission/sideshowbarker/script-off...</li>
 <li id="l-124"><a href="#l-124">#</a> [11:05] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-125" _a="MikeSmith"><a href="#l-125">#</a> [11:07] &lt;MikeSmith&gt; I'm just getting all the love today</li>
 <li id="l-126" _a="MikeSmith"><a href="#l-126">#</a> [11:07] * MikeSmith wipes away a few tears of joy</li>
 <li id="l-127" _a="Ms2ger" _r="MikeSmith"><a href="#l-127">#</a> [11:16] &lt;Ms2ger&gt; MikeSmith, must've done something useful for once ;)</li>
 <li id="l-128" _a="MikeSmith" _r="Ms2ger"><a href="#l-128">#</a> [11:18] &lt;MikeSmith&gt; Ms2ger: sticks and stones -- you can't make me feel sad, because I finally found a place where I can feel welcome <a href="http://www.woodcountycowboychurch.org">http://www.woodcountycowboychurch.org</a></li>
 <li id="l-129" _a="MikeSmith"><a href="#l-129">#</a> [11:18] &lt;MikeSmith&gt; &quot;Where it's OK to NOT be OK!&quot;</li>
 <li id="l-130" _a="MikeSmith"><a href="#l-130">#</a> [11:19] &lt;MikeSmith&gt; there's even some videos of me leading the music worship <a href="https://twitter.com/Woodcountycc/status/498668547501850624">https://twitter.com/Woodcountycc/status/498668547501850624</a></li>
 <li id="l-131"><a href="#l-131">#</a> [11:44] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-132"><a href="#l-132">#</a> [11:51] * Joins: shepazu (schepers@public.cloak)</li>
 <li id="l-133"><a href="#l-133">#</a> [11:51] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-134"><a href="#l-134">#</a> [11:58] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-135"><a href="#l-135">#</a> [11:58] -gitbot:#testing- [web-platform-tests] Ms2ger opened pull request #1207: Add tests for null/undefined/empty string namespace arguments to Document.createElementNS. (master...createElementNS-null-ns) <a href="https://github.com/w3c/web-platform-tests/pull/1207">https://github.com/w3c/web-platform-tests/pull/1207</a></li>
 <li id="l-136"><a href="#l-136">#</a> [11:58] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-137" _a="Ms2ger" _r="MikeSmith"><a href="#l-137">#</a> [11:58] &lt;Ms2ger&gt; MikeSmith, r?</li>
 <li id="l-138" _a="MikeSmith" _r="Ms2ger"><a href="#l-138">#</a> [12:05] &lt;MikeSmith&gt; Ms2ger: On the train but can check it when I get home</li>
 <li id="l-139" _a="MikeSmith"><a href="#l-139">#</a> [12:05] &lt;MikeSmith&gt; or try reviewing from my mobile</li>
 <li id="l-140"><a href="#l-140">#</a> [12:39] * Quits: ptressel (~chatzilla@public.cloak) (&quot;zzz&quot;)</li>
 <li id="l-141"><a href="#l-141">#</a> [12:45] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-142"><a href="#l-142">#</a> [12:51] * Joins: glenn_ (~gadams@public.cloak)</li>
 <li id="l-143"><a href="#l-143">#</a> [12:52] * Quits: glenn (~gadams@public.cloak) (Client closed connection)</li>
 <li id="l-144"><a href="#l-144">#</a> [13:07] * Quits: glenn_ (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-145" _a="Ms2ger" _r="zcorpan_"><a href="#l-145">#</a> [13:38] &lt;Ms2ger&gt; zcorpan_, r? <a href="https://critic.hoppipolla.co.uk/r/2421">https://critic.hoppipolla.co.uk/r/2421</a></li>
 <li id="l-146" _a="zcorpan_"><a href="#l-146">#</a> [13:39] &lt;zcorpan_&gt; r+</li>
 <li id="l-147" _a="Ms2ger"><a href="#l-147">#</a> [13:39] &lt;Ms2ger&gt; Ta</li>
 <li id="l-148" _a="Ms2ger" _r="MikeSmith"><a href="#l-148">#</a> [13:39] &lt;Ms2ger&gt; MikeSmith, un-r?</li>
 <li id="l-149"><a href="#l-149">#</a> [13:41] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-150"><a href="#l-150">#</a> [13:41] -gitbot:#testing- [web-platform-tests] Ms2ger closed pull request #1207: Add tests for null/undefined/empty string namespace arguments to Document.createElementNS. (master...createElementNS-null-ns) <a href="https://github.com/w3c/web-platform-tests/pull/1207">https://github.com/w3c/web-platform-tests/pull/1207</a></li>
 <li id="l-151"><a href="#l-151">#</a> [13:41] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-152" _a="heycam"><a href="#l-152">#</a> [13:44] * heycam is now known as heycam|away</li>
 <li id="l-153" _a="MikeSmith"><a href="#l-153">#</a> [13:44] &lt;MikeSmith&gt; dang</li>
 <li id="l-154" _a="MikeSmith"><a href="#l-154">#</a> [13:44] &lt;MikeSmith&gt; I just got home</li>
 <li id="l-155" _a="MikeSmith"><a href="#l-155">#</a> [13:44] &lt;MikeSmith&gt; ah well</li>
 <li id="l-156" _a="MikeSmith"><a href="#l-156">#</a> [13:44] &lt;MikeSmith&gt; ping me on the next one</li>
 <li id="l-157"><a href="#l-157">#</a> [13:59] * Joins: plh (plehegar@public.cloak)</li>
 <li id="l-158" _a="Ms2ger"><a href="#l-158">#</a> [13:59] &lt;Ms2ger&gt; Will do :)</li>
 <li id="l-159"><a href="#l-159">#</a> [14:01] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-160"><a href="#l-160">#</a> [14:08] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-161"><a href="#l-161">#</a> [14:10] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-162"><a href="#l-162">#</a> [14:10] -gitbot:#testing- [web-platform-tests] Ms2ger opened pull request #1208: Move the data from interfaces.js into a separate JS file. (master...elements) <a href="https://github.com/w3c/web-platform-tests/pull/1208">https://github.com/w3c/web-platform-tests/pull/1208</a></li>
 <li id="l-163"><a href="#l-163">#</a> [14:10] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-164" _a="Ms2ger" _r="MikeSmith"><a href="#l-164">#</a> [14:10] &lt;Ms2ger&gt; MikeSmith, ping :)</li>
 <li id="l-165" _a="MikeSmith"><a href="#l-165">#</a> [14:27] &lt;MikeSmith&gt; taking a look now</li>
 <li id="l-166" _a="Ms2ger"><a href="#l-166">#</a> [14:30] * Ms2ger disappears for most of the rest of the day</li>
 <li id="l-167" _a="MikeSmith" _r="Ms2ger"><a href="#l-167">#</a> [14:32] &lt;MikeSmith&gt; Ms2ger: interfaces.js file not checked in, it seems</li>
 <li id="l-168"><a href="#l-168">#</a> [14:37] * Quits: Ms2ger (~Ms2ger@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-169"><a href="#l-169">#</a> [14:47] * Quits: zcorpan_ (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-170" _a="heycam|away"><a href="#l-170">#</a> [15:00] * heycam|away is now known as heycam</li>
 <li id="l-171" _a="ato"><a href="#l-171">#</a> [15:01] &lt;ato&gt; What constitutes a testharness test?  I.e. what do I need for it to be picked up tools/scripts/manifest.py?</li>
 <li id="l-172" _a="jgraham"><a href="#l-172">#</a> [15:01] &lt;jgraham&gt; &lt;script src=&quot;/resources/testharness.js&quot;&gt;&lt;/script&gt;</li>
 <li id="l-173" _a="jgraham"><a href="#l-173">#</a> [15:01] &lt;jgraham&gt; (an element, not just that string)</li>
 <li id="l-174"><a href="#l-174">#</a> [15:02] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-175" _a="ato"><a href="#l-175">#</a> [15:03] &lt;ato&gt; Done that, but it's not getting picked up…</li>
 <li id="l-176" _a="ato"><a href="#l-176">#</a> [15:03] &lt;ato&gt; Put it in a new subdirectory from the root directory in a file called test_name.html.</li>
 <li id="l-177"><a href="#l-177">#</a> [15:09] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-178" _a="ato"><a href="#l-178">#</a> [15:11] &lt;ato&gt; Does the file have to be checked in to git?</li>
 <li id="l-179" _a="jgraham" _r="Yeah"><a href="#l-179">#</a> [15:12] &lt;jgraham&gt; Yeah, unless you pass the option to include local changes</li>
 <li id="l-180" _a="ato"><a href="#l-180">#</a> [15:13] &lt;ato&gt; Yeah I did that too, -rc.</li>
 <li id="l-181" _a="ato"><a href="#l-181">#</a> [15:30] &lt;ato&gt; It seems the problem is that it doesn't pick up new directories even if -rc is passed, but putting it in an existing directory works.</li>
 <li id="l-182"><a href="#l-182">#</a> [15:38] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-183" _a="jgraham" _r="Yeah"><a href="#l-183">#</a> [15:44] &lt;jgraham&gt; Yeah, so I have a rewrite of manifest.py kicking around that might help. You could try it out if you want</li>
 <li id="l-184"><a href="#l-184">#</a> [15:51] * Joins: AutomatedTester (~AutomatedTester@public.cloak)</li>
 <li id="l-185"><a href="#l-185">#</a> [15:54] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-186"><a href="#l-186">#</a> [15:54] -gitbot:#testing- [web-platform-tests] sideshowbarker closed pull request #1077: w3c-test.org cleanup in html/dnd (master...plh/html-dnd-links) <a href="https://github.com/w3c/web-platform-tests/pull/1077">https://github.com/w3c/web-platform-tests/pull/1077</a></li>
 <li id="l-187"><a href="#l-187">#</a> [15:54] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-188"><a href="#l-188">#</a> [15:56] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-189"><a href="#l-189">#</a> [15:56] * Joins: anssik (~uid10742@public.cloak)</li>
 <li id="l-190"><a href="#l-190">#</a> [16:01] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-191"><a href="#l-191">#</a> [16:01] -gitbot:#testing- [web-platform-tests] jgraham created jgraham/manifest_update_rewrite (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/4d458f9083d3">https://github.com/w3c/web-platform-tests/commit/4d458f9083d3</a></li>
 <li id="l-192"><a href="#l-192">#</a> [16:01] -gitbot:#testing- web-platform-tests/jgraham/manifest_update_rewrite 4d458f9 James Graham: Rewrite manifest.py to make including local changes safer (not very backwards compatible)</li>
 <li id="l-193"><a href="#l-193">#</a> [16:01] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-194" _a="jgraham" _r="ato"><a href="#l-194">#</a> [16:02] &lt;jgraham&gt; ato: <a href="https://github.com/w3c/web-platform-tests/tree/jgraham/manifest_update_rewrite">https://github.com/w3c/web-platform-tests/tree/jgraham/manifest_update_rewrite</a></li>
 <li id="l-195" _a="ato"><a href="#l-195">#</a> [16:07] &lt;ato&gt; Also files in the root directory aren't picked up.</li>
 <li id="l-196"><a href="#l-196">#</a> [16:07] * Quits: scott_gonzalez (~scott_gonzalez@public.cloak) (scott_gonzalez)</li>
 <li id="l-197" _a="ato"><a href="#l-197">#</a> [16:07] &lt;ato&gt; Haven't tested your patch yet.</li>
 <li id="l-198" _a="jgraham" _r="Yeah"><a href="#l-198">#</a> [16:08] &lt;jgraham&gt; Yeah, you shouldn't have tests in the root directory, so that might even be deliberate</li>
 <li id="l-199" _a="ato"><a href="#l-199">#</a> [16:09] &lt;ato&gt; Any idea where I can put a unit test for a tool which needs to be picked up by wptrunner?</li>
 <li id="l-200" _a="ato"><a href="#l-200">#</a> [16:10] &lt;ato&gt; I see resources/ is also on the blacklist.</li>
 <li id="l-201"><a href="#l-201">#</a> [16:39] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-202"><a href="#l-202">#</a> [16:39] -gitbot:#testing- [web-platform-tests] sideshowbarker closed pull request #1078: Prevent the use of w3c-test.org in the tests (master...plh/w3ctestorgrule) <a href="https://github.com/w3c/web-platform-tests/pull/1078">https://github.com/w3c/web-platform-tests/pull/1078</a></li>
 <li id="l-203"><a href="#l-203">#</a> [16:39] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-204" _a="plh"><a href="#l-204">#</a> [17:10] &lt;plh&gt; yoohoo! no more w3c-test.org in the tests</li>
 <li id="l-205" _a="jgraham"><a href="#l-205">#</a> [17:11] &lt;jgraham&gt; Great!</li>
 <li id="l-206"><a href="#l-206">#</a> [17:14] * Joins: scott_gonzalez (~scott_gonzalez@public.cloak)</li>
 <li id="l-207"><a href="#l-207">#</a> [17:15] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
 <li id="l-208"><a href="#l-208">#</a> [17:17] * Quits: dom (dom@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-209"><a href="#l-209">#</a> [17:39] * Quits: AutomatedTester (~AutomatedTester@public.cloak) (Client closed connection)</li>
 <li id="l-210" _a="Ms2ger" _r="jgraham"><a href="#l-210">#</a> [17:42] &lt;Ms2ger&gt; jgraham, r? <a href="https://critic.hoppipolla.co.uk/r/2424">https://critic.hoppipolla.co.uk/r/2424</a></li>
 <li id="l-211" _a="heycam"><a href="#l-211">#</a> [17:50] * heycam is now known as heycam|away</li>
 <li id="l-212" _a="plh" _r="Ms2ger"><a href="#l-212">#</a> [17:52] &lt;plh&gt; Ms2ger, you wrote assert_true(desc.writable, &quot;property is not writable&quot;);</li>
 <li id="l-213" _a="plh"><a href="#l-213">#</a> [17:53] &lt;plh&gt; did you mean to say property is writable?</li>
 <li id="l-214" _a="plh" _r="hum"><a href="#l-214">#</a> [17:53] &lt;plh&gt; hum, I might get my negation wrong</li>
 <li id="l-215" _a="Ms2ger" _r="plh"><a href="#l-215">#</a> [17:53] &lt;Ms2ger&gt; plh, that's the error message</li>
 <li id="l-216" _a="plh"><a href="#l-216">#</a> [17:53] &lt;plh&gt; ok</li>
 <li id="l-217" _a="ato" _r="jgraham"><a href="#l-217">#</a> [17:55] &lt;ato&gt; jgraham:  0:09.16 LOG: Thread-TestrunnerManager-1 WARNING Failed to find base prefs file in /Library/Python/2.7/site-packages/wptrunner/prefs/prefs_general.js</li>
 <li id="l-218" _a="heycam|away"><a href="#l-218">#</a> [17:59] * heycam|away is now known as heycam</li>
 <li id="l-219" _a="ato"><a href="#l-219">#</a> [18:02] &lt;ato&gt; Hm I just upgraded wptrunner, and the following command isn't working anymore: `wptrunner --binary /Applications/FirefoxNightly.app/Contents/MacOS/firefox --product firefox --processes 1 --log-mach - . . --include tools`</li>
 <li id="l-220" _a="ato"><a href="#l-220">#</a> [18:05] &lt;ato&gt; Ah.  It's `--tests . --metadata .` now.</li>
 <li id="l-221"><a href="#l-221">#</a> [18:05] * Quits: mdas (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-222"><a href="#l-222">#</a> [18:05] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-223" _a="ato" _r="jgraham"><a href="#l-223">#</a> [18:08] &lt;ato&gt; jgraham: What does this mean?  0:09.33 LOG: Thread-TestrunnerManager-1 DEBUG Unexpected count in this thread 1</li>
 <li id="l-224" _a="ato" _r="jgraham"><a href="#l-224">#</a> [18:08] &lt;ato&gt; jgraham: Also it's claiming that my test crashes.  How can I debug that?</li>
 <li id="l-225" _a="Ms2ger" _r="ato"><a href="#l-225">#</a> [18:09] &lt;Ms2ger&gt; ato, that there's one unexpected result on that thread :)</li>
 <li id="l-226"><a href="#l-226">#</a> [18:12] * Quits: mdas (~mdas@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-227" _a="jgraham" _r="Ms2ger"><a href="#l-227">#</a> [18:14] &lt;jgraham&gt; Ms2ger: FWIW I find using error messages like that confusing</li>
 <li id="l-228" _a="jgraham"><a href="#l-228">#</a> [18:15] &lt;jgraham&gt; I think &quot;property should be writable&quot;</li>
 <li id="l-229" _a="Ms2ger"><a href="#l-229">#</a> [18:15] &lt;Ms2ger&gt; Fair</li>
 <li id="l-230" _a="Ms2ger"><a href="#l-230">#</a> [18:15] &lt;Ms2ger&gt; I stole them all from Aryeh, though</li>
 <li id="l-231"><a href="#l-231">#</a> [18:15] * Quits: scott_gonzalez (~scott_gonzalez@public.cloak) (scott_gonzalez)</li>
 <li id="l-232"><a href="#l-232">#</a> [18:15] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-233" _a="jgraham"><a href="#l-233">#</a> [18:16] &lt;jgraham&gt; (that way it ends up as &quot;assert_true: got false, property should be writable&quot; rather than &quot;assert_true: got false, property is not writable&quot;)</li>
 <li id="l-234" _a="jgraham" _r="Yeah"><a href="#l-234">#</a> [18:16] &lt;jgraham&gt; Yeah, I assumed as much</li>
 <li id="l-235" _a="jgraham" _r="ato"><a href="#l-235">#</a> [18:16] &lt;jgraham&gt; ato: Well debug how?</li>
 <li id="l-236" _a="ato"><a href="#l-236">#</a> [18:16] &lt;ato&gt; Well it turns out I was using the wrong executor.</li>
 <li id="l-237" _a="Ms2ger" _r="jgraham"><a href="#l-237">#</a> [18:17] &lt;Ms2ger&gt; jgraham, I'd be happy to try to search/replace some of those, but not right now</li>
 <li id="l-238" _a="jgraham" _r="Ms2ger"><a href="#l-238">#</a> [18:18] &lt;jgraham&gt; Ms2ger: Fair enough, I wasn't really expecting you to</li>
 <li id="l-239" _a="heycam"><a href="#l-239">#</a> [18:23] * heycam is now known as heycam|away</li>
 <li id="l-240"><a href="#l-240">#</a> [18:40] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-241"><a href="#l-241">#</a> [18:43] * Quits: anssik (~uid10742@public.cloak) (&quot;Connection closed for inactivity&quot;)</li>
 <li id="l-242"><a href="#l-242">#</a> [18:49] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-243"><a href="#l-243">#</a> [18:49] -gitbot:#testing- [web-platform-tests] Ms2ger opened pull request #1209: Update interfaces tests. (master...update-interfaces) <a href="https://github.com/w3c/web-platform-tests/pull/1209">https://github.com/w3c/web-platform-tests/pull/1209</a></li>
 <li id="l-244"><a href="#l-244">#</a> [18:49] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-245" _a="Ms2ger"><a href="#l-245">#</a> [18:50] &lt;Ms2ger&gt; rs?</li>
 <li id="l-246"><a href="#l-246">#</a> [19:01] * Joins: anssik (~uid10742@public.cloak)</li>
 <li id="l-247"><a href="#l-247">#</a> [19:23] * Quits: Ms2ger (~Ms2ger@public.cloak) (&quot;Leaving&quot;)</li>
 <li id="l-248"><a href="#l-248">#</a> [19:33] * Joins: mdas_ (~mdas@public.cloak)</li>
 <li id="l-249"><a href="#l-249">#</a> [19:33] * Quits: mdas (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-250"><a href="#l-250">#</a> [20:16] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-251"><a href="#l-251">#</a> [20:24] * Joins: scott_gonzalez (~scott_gonzalez@public.cloak)</li>
 <li id="l-252"><a href="#l-252">#</a> [20:47] * Joins: ptressel (~chatzilla@public.cloak)</li>
 <li id="l-253"><a href="#l-253">#</a> [20:48] * Joins: AutomatedTester (~AutomatedTester@public.cloak)</li>
 <li id="l-254"><a href="#l-254">#</a> [21:13] * Quits: anssik (~uid10742@public.cloak) (&quot;Connection closed for inactivity&quot;)</li>
 <li id="l-255"><a href="#l-255">#</a> [21:23] * Quits: ptressel (~chatzilla@public.cloak) (&quot;bbl&quot;)</li>
 <li id="l-256"><a href="#l-256">#</a> [21:31] * Quits: mdas_ (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-257"><a href="#l-257">#</a> [21:31] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-258"><a href="#l-258">#</a> [22:24] * Quits: plh (plehegar@public.cloak) (&quot;Leaving&quot;)</li>
 <li id="l-259"><a href="#l-259">#</a> [22:35] * Joins: MK (kawamori@public.cloak)</li>
 <li id="l-260"><a href="#l-260">#</a> [22:52] * Quits: mdas (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-261"><a href="#l-261">#</a> [22:52] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-262"><a href="#l-262">#</a> [23:12] * Quits: mdas (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-263"><a href="#l-263">#</a> [23:13] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-264"><a href="#l-264">#</a> [23:14] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
 <li id="l-265"><a href="#l-265">#</a> [23:20] * Quits: mdas (~mdas@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-266"><a href="#l-266">#</a> Session Close: Fri Aug 22 00:00:00 2014</li>
</ol>
<p id="end">The end :)
